+++ iceoryx_utils/source/posix_wrapper/shared_memory_object/allocator.cpp
@@ -42,9 +42,9 @@ void* Allocator::allocate(const uint64_t f_size, const uint64_t f_alignment)
         std::terminate();
     }
 
-    uintptr_t l_currentAddress = reinterpret_cast<uintptr_t>(m_startAddress) + m_currentPosition;
-    uintptr_t l_alignedPosition = cxx::align(l_currentAddress, static_cast<uintptr_t>(f_alignment));
-    l_alignedPosition -= reinterpret_cast<uintptr_t>(m_startAddress);
+    uint64_t l_currentAddress = reinterpret_cast<uint64_t>(m_startAddress) + m_currentPosition;
+    uint64_t l_alignedPosition = cxx::align(l_currentAddress, static_cast<uint64_t>(f_alignment));
+    l_alignedPosition -= reinterpret_cast<uint64_t>(m_startAddress);
 
     byte_t* l_returnValue = nullptr;
